+++ CHANGELOG.md
@@ -1,0 +1,5 @@
+## v5.0.0 Breaking changes
+
+- `calculations`: Removed `getRadian` export
+- `data`: Removed `setField` export, use [`lodash.set(object, path, value)`](https://lodash.com/docs/4.17.15#set) instead
+- `formatters`: Remove `capitalizeFirst` export, use [`lodash.upperFirst(string)`](https://lodash.com/docs/4.17.15#upperFirst) instead
