+++ package.json
@@ -7,7 +7,7 @@
   "homepage": "https://nebenan.de/",
   "repository": "good-hood-gmbh/nebenan-helpers",
   "bugs": "https://github.com/good-hood-gmbh/nebenan-helpers/issues",
-  "version": "4.7.1",
+  "version": "5.0.0",
   "files": [
     "lib/*.js",
     "lib/*/*.js"
