+++ src/calculations/utils.js
@@ -1,0 +1,1 @@
+export const getRadian = (degrees) => (degrees * (Math.PI / 180));
